requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
pytz==2026.1.post1
qrcode==8.2
sqlparse==0.5.5
tzdata==2025.3
razorpay
